Add files via upload
Welcome to the ToDo List Website repository, a dynamic and visually appealing task management solution built with the trifecta of web technologies: HTML, CSS, and JavaScript. This project is designed to help you organize your tasks efficiently and can serve as a foundation for building feature-rich to-do list applications.  The ToDo List Website boasts a clean and user-friendly interface that makes managing tasks a breeze. Easily add, edit, and remove tasks with a few clicks.  The dynamic behavior of the ToDo List is powered by JavaScript, allowing for real-time updates without the need for page reloads. The code is well-organized and commented for easy understanding and modification. Add tasks with due dates, mark them as completed, and remove them when no longer needed. The ToDo List Website provides essential features to keep your tasks organized.  User tasks are stored locally, ensuring that the ToDo list persists even if the user refreshes the page or closes the browser. This feature enhances user experience by retaining task data across sessions.